Praveen-feat: Created new Attendees and Agenda models (#30)
* Added search by name or description endpoint

* Created Attendee, Agenda, Agendasession tables

* Added schema files for attendee,agenda,agendasessions

* Added Attendee end points and Attendee CRUD

* Added Agenda End points and Agenda CRUD

* Made User endpoints wokring with the user_id from the token

* Modified User Model to support Authorization

* Updated Users Model

* Updated Conference Model, Conference CRUD to support authorization

* Updated Sessions Model, Sessions CRUD to support authorization

* Updated Settings Models, Settings Crud to support authentication

* Updated Session endpoint paths

* Updated requirements file

* Removed dummy file

* Date range bug fix

* Updated sessions search filter

* removed stale imports

* Updated search filter

* Update sql_app.db

* Updated Conference date range amd search sessions filter

* Updated Sessions search

* sessions update

* Updated User Model to return encrypted ID

* Updated User Model to return encrypted ID

* Updated Encryption for Conference_Id

* Updated Sessions for conference id encryption

* Updated Settings for Id encryption

* Updated Users Model

* Updated Users Model

* Updated Conferences

* Updated Sessions Model for encryption

* Updated Settings model for Id encryption

* OTP generation

* DB Commit

* Updated atandee

* Db Commit

* Remove encryption of session IDs and owner IDs

* Remove encryption of conference IDs and owner IDs

* Delete encryption.py file.

* Remove encryption of settings IDs and owner IDs

* Remove unused import and function call in users.py

* Delete sql_app.db file

* Add UUID to attendee creation

* Fix typo in hashed_password field and add
validators to attendee schemas

* Add validation for empty request body in update
endpoints

* Add valid_email variable to password_reset
function

* Fix OTP generator initialization and reset after
password reset

* Refactor routers and schemas for better
organization and readability

* Add delete_agenda_by_conference_id function and
update delete_conference and delete_session
functions

* Normalize email addresses in attendee and user
routers

* Update user password validation and error handling

* Add company field to user model and update user
CRUD operations

+++ app/crud/conferences_crud.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session
 from datetime import datetime, date
 from .. import models
 from ..schemas import conference_schemas as schemas
+from . import agenda_crud
 from pytz import timezone
 import uuid
 
@@ -27,14 +28,6 @@ def create_user_conference(db: Session, conference: schemas.ConferenceCreate, us
     db.refresh(db_conference)
     return db_conference
 
-# get conference by conference id
-def get_conference(db: Session, conference_id: int):
-    return db.query(models.Conference).filter(models.Conference.id == conference_id).first()
-
-#get conference by owner id and conference id
-def get_conference_by_uuid_id(db: Session, owner_id: int, uuid: int):
-    return db.query(models.Conference).filter(models.Conference.owner_id == owner_id, models.Conference.uuid == uuid).first()
-
 # get conference by uuid and owner id
 def get_conference_by_uuid(db: Session, uuid: str, owner_id: int):
     return db.query(models.Conference).filter(models.Conference.uuid == uuid, models.Conference.owner_id == owner_id).first()
@@ -43,11 +36,16 @@ def get_conference_by_conference_uuid(db: Session, uuid: str):
     return db.query(models.Conference).filter(models.Conference.uuid == uuid).first()
 
 # delete conference by conference id
-def delete_conference(db: Session,owner_id: int, uuid: str):
+def delete_conference(db: Session, owner_id: int, uuid: str):
     conference = db.query(models.Conference).filter(models.Conference.uuid == uuid, models.Conference.owner_id == owner_id).first()
-    db.query(models.Session).filter(models.Session.conference_id == conference.id, models.Session.owner_id==owner_id).delete()
-    db.query(models.Settings).filter(models.Settings.conference_id == conference.id, models.Settings.owner_id==owner_id).delete()
-    db.query(models.Conference).filter(models.Conference.uuid == uuid, models.Conference.owner_id == owner_id).delete()
+    if conference is None:
+        return False
+    sessions = db.query(models.Session).filter(models.Session.conference_id == conference.id, models.Session.owner_id == owner_id)
+    for session in sessions:
+        db.delete(session)
+    db.query(models.Settings).filter(models.Settings.conference_id == conference.id, models.Settings.owner_id == owner_id).delete()
+    agenda_crud.delete_agenda_by_conference_id(db, conference_id=conference.id)
+    db.delete(conference)
     db.commit()
     return True
